feat: make all rules `error` instead of `warn` (#356)
Originally when I authored our configs I set a handful of rules to
`warn` instead of `error` as a way of making the configs gentler and
hint to people when it might be ok to actually disable the rule for that
particular violation depending on the context.

In practice however, most people tend to expect the linter to block in
CI and that if it doesn't then everything is fine, meaning the rules
configured to `warn` were effectively the same as not being enabled at
all except for when people run `--fix`.

When we've talked about this recently, the general consensus was people
would prefer the rules just `error`, and so that's what this does